Vase Setup for MIssionTutorial + Naming Refactors
