+++ Assets/Prefabs/Missions/MissionTutorial.prefab
@@ -45,14 +45,14 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   playerStartPosition: {x: -95, y: 5, z: 135}
   vase:
-    prefab: {fileID: 3658902531811208850, guid: 90321af8d099f0b43850159f7ad318f7,
+    prefab: {fileID: 6872363483067291833, guid: 23f56f3b92555d04a801b412248a65f1,
       type: 3}
-    position: {x: -69, y: 5, z: 100}
+    position: {x: -67, y: 3, z: 103}
     rotation: {x: 0, y: 0, z: 0}
     spawnedInstance: {fileID: 0}
   note:
-    prefab: {fileID: 3658902531811208850, guid: 90321af8d099f0b43850159f7ad318f7,
+    prefab: {fileID: 9073910147309779605, guid: 46116c0641b9a6a44845879c14655d30,
       type: 3}
-    position: {x: -69, y: 2, z: 135}
+    position: {x: -68, y: 1.5, z: 102}
     rotation: {x: 0, y: 0, z: 0}
     spawnedInstance: {fileID: 0}
