+++ Assets/Prefabs/Props/Vase_Stand_Temp.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 46116c0641b9a6a44845879c14655d30
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
