+++ Assets/Prefabs/Props/Vase_Uncut.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 23f56f3b92555d04a801b412248a65f1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
