+++ Assets/Resources/Generic_Models/VaseMaterials/vaseMatShake.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e9d88d1afd1a61b49b58dcd48dd19e19
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
