+++ Assets/Scripts/Editor/MissionCheats.cs
@@ -36,28 +36,28 @@ public class MissionCheats: MonoBehaviour
     ////////////////////
     // KITCHEN LEVEL
     ////////////////////
-    [MenuItem(Constants.CHEATS_STARTMISSIONCAFETERIA1, true)]
+    [MenuItem(Constants.CHEATS_STARTMISSIONKITCHEN1, true)]
     public static bool ValidateStartMissionCafeteria1()
     {
-        return Application.isPlaying && MissionManager.Instance && MissionManager.Instance.GetActiveMission<MissionCafeteria1>() == null;
+        return Application.isPlaying && MissionManager.Instance && MissionManager.Instance.GetActiveMission<MissionKitchen1>() == null;
     }
 
-    [MenuItem(Constants.CHEATS_STARTMISSIONCAFETERIA1)]
+    [MenuItem(Constants.CHEATS_STARTMISSIONKITCHEN1)]
     public static void StartMissionCafeteria1()
     {
         MissionManager.Instance.StartMission(AssetDatabase.LoadAssetAtPath<GameObject>(Constants.ASSET_PATH_MISSIONCAFETERIA1));
     }
 
-    [MenuItem(Constants.CHEATS_ENDMISSIONCAFETERIA1, true)]
+    [MenuItem(Constants.CHEATS_ENDMISSIONKITCHEN1, true)]
     public static bool ValidateStopMissionCafeteria1()
     {
-        return Application.isPlaying && MissionManager.Instance && MissionManager.Instance.GetActiveMission<MissionCafeteria1>() != null;
+        return Application.isPlaying && MissionManager.Instance && MissionManager.Instance.GetActiveMission<MissionKitchen1>() != null;
     }
 
-    [MenuItem(Constants.CHEATS_ENDMISSIONCAFETERIA1)]
+    [MenuItem(Constants.CHEATS_ENDMISSIONKITCHEN1)]
     public static void EndMissionCafeteria1()
     {
-        AMission mission = MissionManager.Instance.GetActiveMission<MissionCafeteria1>().mission;
+        AMission mission = MissionManager.Instance.GetActiveMission<MissionKitchen1>().mission;
         MissionManager.Instance.EndMission(mission);
     }
 }
