+++ Assets/Scripts/Interactables/BreakableObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bdf03f57578a27345be3f6441652e25b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
