+++ Assets/Scripts/LoudObjects/LoudObject.cs
@@ -21,6 +21,9 @@ public class LoudObject : MonoBehaviour
     private bool hasHit = false;
     private NoisePing noisePing;
 
+    public delegate void HasHitAction();
+    public event HasHitAction OnHit;
+
     void Start()
     {
         rb = GetComponent<Rigidbody>();
@@ -53,6 +56,8 @@ public class LoudObject : MonoBehaviour
 
         noisePing.SpawnNoisePing(other);
         hasHit = true;
+
+        OnHit?.Invoke();
     }
 
     void Shake(float shake){
@@ -71,9 +76,13 @@ public class LoudObject : MonoBehaviour
         Vector3 directionToDraw = transform.TransformDirection(thrustDirection) * 5;
         Gizmos.DrawRay(transform.position, directionToDraw);
 
-        // Radius sphere 
+        // Shake radius sphere 
         Gizmos.color = Color.white;
         Gizmos.DrawWireSphere(transform.position, shakeRadius);
+        
+        // Drop radius sphere 
+        Gizmos.color = Color.blue;
+        Gizmos.DrawWireSphere(transform.position, dropRadius);
     }
 
 }
