--- Assets/Scripts/Missions/MissionCafeteria1.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 54ff655fb6abc834888711c127219b0f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
