+++ Assets/Scripts/Missions/MissionKitchen1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afb112b1dbeb55f408e02b5b606efadb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
